updated orders#create to factor in product.available and updated order item model to be dependeent destroy on order

+++ app/controllers/orders_controller.rb
@@ -1,14 +1,24 @@
 class OrdersController < ApplicationController
     def create
-        @order = Order.create!(user_id: current_user.id)
-        current_user.cart.cart_items.each do |cart_item|
-            OrderItem.create!(order_id: @order.id, product_id: cart_item.product.id)
-            cart_item.product.update(available: false)
-            cart_item.destroy
-        end     
-        @order_items = @order.order_items
-        flash.notice = "Order placed"
-        redirect_to show_user_cart_path
+        if current_user.cart.cart_items.all? { |cart_item| cart_item.product.available }
+            @order = Order.create!(user_id: current_user.id)
+            current_user.cart.cart_items.each do |cart_item|
+                OrderItem.create!(order_id: @order.id, product_id: cart_item.product.id)
+                cart_item.product.update(available: false)
+                cart_item.destroy
+            end     
+            @order_items = @order.order_items
+            flash.notice = "Order placed"
+            redirect_to show_user_cart_path
+        else
+            current_user.cart.cart_items.each do |cart_item|
+                if !cart_item.product.available
+                    cart_item.destroy
+                end
+            end
+            flash.alert = "One or more items unavailable"
+            redirect_to show_user_cart_path
+        end
     end
 
     def show
